+++ src/server/rest/service/SiteService.ts
@@ -201,6 +201,17 @@ export default class SiteService {
     const site = await SiteStorage.getSite(req.user.tenantID, filteredRequest.siteID);
     UtilsService.assertObjectExists(action, site, `Site '${filteredRequest.siteID}' does not exist`,
       MODULE_NAME, 'handleAssignUsersToSite', req.user);
+    // OCPI Site
+    if (!site.issuer) {
+      throw new AppError({
+        source: Constants.CENTRAL_SERVER,
+        errorCode: HTTPError.GENERAL_ERROR,
+        message: `Site '${site.name}' with ID '${site.id}' not issued by the organization`,
+        module: MODULE_NAME, method: 'handleAssignUsersToSite',
+        user: req.user,
+        action: action
+      });
+    }
     // Get Users
     for (const userID of filteredRequest.userIDs) {
       // Check the user
@@ -215,6 +226,17 @@ export default class SiteService {
           action: Action.READ, entity: Entity.USER,
           module: MODULE_NAME, method: 'handleAssignUsersToSite',
           value: userID
+        });
+      }
+      // OCPI User
+      if (!user.issuer) {
+        throw new AppError({
+          source: Constants.CENTRAL_SERVER,
+          errorCode: HTTPError.GENERAL_ERROR,
+          message: `User '${Utils.buildUserFullName(user)}' not issued by the organization`,
+          module: MODULE_NAME, method: 'handleAssignUsersToSite',
+          user: req.user, actionOnUser: user,
+          action: action
         });
       }
     }
@@ -313,6 +335,17 @@ export default class SiteService {
     const site = await SiteStorage.getSite(req.user.tenantID, siteID);
     UtilsService.assertObjectExists(action, site, `Site with ID '${siteID}' does not exist`,
       MODULE_NAME, 'handleDeleteSite', req.user);
+    // OCPI Site
+    if (!site.issuer) {
+      throw new AppError({
+        source: Constants.CENTRAL_SERVER,
+        errorCode: HTTPError.GENERAL_ERROR,
+        message: `Site '${site.name}' with ID '${site.id}' not issued by the organization`,
+        module: MODULE_NAME, method: 'handleDeleteSite',
+        user: req.user,
+        action: action
+      });
+    }
     // Delete
     await SiteStorage.deleteSite(req.user.tenantID, site.id);
     // Log
@@ -453,6 +486,17 @@ export default class SiteService {
     const company = await CompanyStorage.getCompany(req.user.tenantID, filteredRequest.companyID);
     UtilsService.assertObjectExists(action, company, `Company ID '${filteredRequest.companyID}' does not exist`,
       MODULE_NAME, 'handleCreateSite', req.user);
+    // OCPI Company
+    if (!company.issuer) {
+      throw new AppError({
+        source: Constants.CENTRAL_SERVER,
+        errorCode: HTTPError.GENERAL_ERROR,
+        message: `Company '${company.name}' with ID '${company.id}' not issued by the organization`,
+        module: MODULE_NAME, method: 'handleCreateSite',
+        user: req.user,
+        action: action
+      });
+    }
     // Create site
     const site: Site = {
       ...filteredRequest,
@@ -497,10 +541,32 @@ export default class SiteService {
     const company = await CompanyStorage.getCompany(req.user.tenantID, filteredRequest.companyID);
     UtilsService.assertObjectExists(action, company, `Company ID '${filteredRequest.companyID}' does not exist`,
       MODULE_NAME, 'handleUpdateSite', req.user);
+    // OCPI Company
+    if (!company.issuer) {
+      throw new AppError({
+        source: Constants.CENTRAL_SERVER,
+        errorCode: HTTPError.GENERAL_ERROR,
+        message: `Company '${company.name}' with ID '${company.id}' not issued by the organization`,
+        module: MODULE_NAME, method: 'handleCreateSite',
+        user: req.user,
+        action: action
+      });
+    }
     // Get Site
     const site: Site = await SiteStorage.getSite(req.user.tenantID, filteredRequest.id);
     UtilsService.assertObjectExists(action, site, `Site with ID '${filteredRequest.id}' does not exist`,
       MODULE_NAME, 'handleUpdateSite', req.user);
+    // OCPI Site
+    if (!company.issuer) {
+      throw new AppError({
+        source: Constants.CENTRAL_SERVER,
+        errorCode: HTTPError.GENERAL_ERROR,
+        message: `Site '${site.name}' with ID '${site.id}' not issued by the organization`,
+        module: MODULE_NAME, method: 'handleUpdateSite',
+        user: req.user,
+        action: action
+      });
+    }
     // Update
     site.lastChangedBy = { 'id': req.user.id };
     site.lastChangedOn = new Date();
